+++ AdministrationAPI/Models/EInvoice/EInvoiceRequest.cs
@@ -1,0 +1,23 @@
+using AdministrationAPI.Models.Vendor;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace AdministrationAPI.Models.EInvoice;
+
+public class EInvoiceRequest
+{
+    [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+    [Key, Column(Order = 0)]
+    public int EInvoiceRequestId { get; set; }
+    [ForeignKey("User")]
+    public string UserId { get; set; }
+    public User User { get; set; }
+    [ForeignKey("Vendor")]
+    public int VendorId { get; set; }
+    public Vendor.Vendor Vendor { get; set; }
+    public int Status { get; set; }
+    public string? Param1 { get; set; }
+    public string? Param2 { get; set; }
+    public string? Param3 { get; set; }
+    public string? Param4 { get; set; }
+}
